Refactored image widget
