+++ frontend/src/components/ImageWidget.tsx
@@ -1,25 +1,37 @@
 import React from "react";
-import { ImageWidget } from "../models";
-import ImagePreview from "./ImagePreview";
-import { useImage } from "../hooks";
+import MarkdownRender from "./MarkdownRender";
 
-interface Props {
-  widget: ImageWidget;
-}
+type Props = {
+  title: string;
+  summary: string;
+  file: File | undefined;
+  summaryBelow: boolean;
+  altText: string;
+};
 
-function ImageWidgetComponent(props: Props) {
-  const { showTitle, content } = props.widget;
-  const file = useImage(content.s3Key.raw);
+const ImageWidget = (props: Props) => {
+  const { file, summaryBelow, summary, title, altText } = props;
 
   return (
-    <ImagePreview
-      title={showTitle ? content.title : ""}
-      summary={content.summary ? content.summary : ""}
-      summaryBelow={content.summaryBelow}
-      file={file.file}
-      altText={content.imageAltText}
-    />
+    <div className="preview-container">
+      <h2 className="margin-left-1 margin-top-1">{title}</h2>
+      {!summaryBelow && (
+        <MarkdownRender
+          source={summary}
+          className="margin-left-1 margin-top-0 margin-bottom-3 imageSummaryAbove"
+        />
+      )}
+      <div className="margin-left-1">
+        <img src={file ? URL.createObjectURL(file) : ""} alt={altText}></img>
+      </div>
+      {summaryBelow && (
+        <MarkdownRender
+          source={summary}
+          className="margin-left-1 margin-top-3 margin-bottom-0 imageSummaryBelow"
+        />
+      )}
+    </div>
   );
-}
+};
 
-export default ImageWidgetComponent;
+export default ImageWidget;
