+++ frontend/src/components/WidgetRender.tsx
@@ -7,7 +7,7 @@ import {
   ImageWidget,
   MetricsWidget,
 } from "../models";
-import { useWidgetDataset } from "../hooks";
+import { useWidgetDataset, useImage } from "../hooks";
 import ChartWidgetComponent from "../components/ChartWidget";
 import TableWidgetComponent from "../components/TableWidget";
 import TextWidget from "../components/TextWidget";
@@ -19,13 +19,41 @@ interface Props {
 }
 
 function WidgetRender({ widget }: Props) {
-  const { json, jsonHeaders } = useWidgetDataset(widget);
-
   switch (widget.widgetType) {
     case WidgetType.Text:
       return <TextWidget widget={widget} />;
     case WidgetType.Chart:
       return <ChartWidgetComponent widget={widget as ChartWidget} />;
+    case WidgetType.Table:
+      return <WidgetWithDataset widget={widget} />;
+    case WidgetType.Image:
+      return <WidgetWithImage widget={widget as ImageWidget} />;
+    case WidgetType.Metrics:
+      return <MetricsWidgetComponent widget={widget as MetricsWidget} />;
+    default:
+      return null;
+  }
+}
+
+function WidgetWithImage({ widget }: Props) {
+  const imageWidget = widget as ImageWidget;
+  const content = imageWidget.content;
+  const file = useImage(content.s3Key.raw);
+
+  return (
+    <ImageWidgetComponent
+      title={imageWidget.showTitle ? content.title : ""}
+      summary={content.summary ? content.summary : ""}
+      summaryBelow={content.summaryBelow}
+      file={file.file}
+      altText={content.imageAltText}
+    />
+  );
+}
+
+function WidgetWithDataset({ widget }: Props) {
+  const { json, jsonHeaders } = useWidgetDataset(widget);
+  switch (widget.widgetType) {
     case WidgetType.Table:
       const tableWidget = widget as TableWidget;
       return (
@@ -37,11 +65,6 @@ function WidgetRender({ widget }: Props) {
           summaryBelow={tableWidget.content.summaryBelow}
         />
       );
-
-    case WidgetType.Image:
-      return <ImageWidgetComponent widget={widget as ImageWidget} />;
-    case WidgetType.Metrics:
-      return <MetricsWidgetComponent widget={widget as MetricsWidget} />;
     default:
       return null;
   }
